requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
click==8.1.3
decorator==5.1.1
dj-database-url==1.2.0
Django==4.0.7
django-environ==0.9.0
django-mkdocs==0.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
executing==1.2.0
ghp-import==2.1.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mergedeep==1.3.4
mkdocs==1.4.2
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-testdox==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
pyyaml_env_tag==0.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
traitlets==5.7.0
uritemplate==4.1.1
watchdog==2.2.1
wcwidth==0.2.5
